+++ supertokens_python/recipe/jwt/utils.py
@@ -13,10 +13,11 @@
 # under the License.
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Callable, Union
+from typing import TYPE_CHECKING, Callable, Union, Optional, Any
 
 if TYPE_CHECKING:
     from .interfaces import APIInterface, RecipeInterface
+    from supertokens_python import AppInfo
 
 
 class OverrideConfig:
@@ -51,3 +52,16 @@ def validate_and_normalise_user_input(
         jwt_validity_seconds = 3153600000
 
     return JWTConfig(override, jwt_validity_seconds)
+
+
+async def get_api_domain_or_throw_error(
+    api_domain: Optional[str], app_info: AppInfo, user_context: Any
+):
+    if api_domain is None:
+        if app_info.initial_api_domain_type == "string":
+            api_domain_res = await app_info.api_domain({}, user_context)  # type:ignore
+            api_domain = api_domain_res.get_as_string_dangerous()
+            return api_domain
+        raise Exception(
+            "Please pass api_domain as a string to the function or pass api_domain as string in supertokens.init"
+        )