+++ supertokens_python/recipe/openid/interfaces.py
@@ -41,6 +41,8 @@ class RecipeInterface(ABC):
         payload: Dict[str, Any],
         validity_seconds: Optional[int],
         use_static_signing_key: Optional[bool],
+        issuer_domain: Optional[str],
+        api_domain: Optional[str],
         user_context: Dict[str, Any],
     ) -> Union[CreateJwtOkResult, CreateJwtResultUnsupportedAlgorithm]:
         pass
@@ -51,7 +53,7 @@ class RecipeInterface(ABC):
 
     @abstractmethod
     async def get_open_id_discovery_configuration(
-        self, user_context: Dict[str, Any]
+        self, issuer_domain: Optional[str], user_context: Dict[str, Any]
     ) -> GetOpenIdDiscoveryConfigurationResult:
         pass