+++ supertokens_python/recipe/openid/syncio/__init__.py
@@ -30,11 +30,18 @@ def create_jwt(
     payload: Optional[Dict[str, Any]] = None,
     validity_seconds: Optional[int] = None,
     use_static_signing_key: Optional[bool] = None,
+    issuer_domain: Optional[str] = None,
+    api_domain: Optional[str] = None,
     user_context: Optional[Dict[str, Any]] = None,
 ) -> Union[CreateJwtOkResult, CreateJwtResultUnsupportedAlgorithm]:
     return sync(
         asyncio.create_jwt(
-            payload, validity_seconds, use_static_signing_key, user_context
+            payload,
+            validity_seconds,
+            use_static_signing_key,
+            issuer_domain,
+            api_domain,
+            user_context,
         )
     )
 
@@ -44,6 +51,8 @@ def get_jwks(user_context: Optional[Dict[str, Any]] = None) -> GetJWKSResult:
 
 
 def get_open_id_discovery_configuration(
-    user_context: Optional[Dict[str, Any]] = None
+    issuer_domain: Optional[str] = None, user_context: Optional[Dict[str, Any]] = None
 ) -> GetOpenIdDiscoveryConfigurationResult:
-    return sync(asyncio.get_open_id_discovery_configuration(user_context))
+    return sync(
+        asyncio.get_open_id_discovery_configuration(issuer_domain, user_context)
+    )