+++ supertokens_python/recipe/session/asyncio/__init__.py
@@ -41,7 +41,7 @@ from ..session_request_functions import (
     get_session_from_request,
     refresh_session_in_request,
 )
-from ..utils import get_required_claim_validators
+from ..utils import get_required_claim_validators, get_api_domain_or_throw_error
 
 _T = TypeVar("_T")
 
@@ -81,6 +81,7 @@ async def create_new_session_without_request_response(
     access_token_payload: Union[Dict[str, Any], None] = None,
     session_data_in_database: Union[Dict[str, Any], None] = None,
     disable_anti_csrf: bool = False,
+    api_domain: Optional[str] = None,
     user_context: Union[None, Dict[str, Any]] = None,
 ) -> SessionContainer:
     if user_context is None:
@@ -94,10 +95,12 @@ async def create_new_session_without_request_response(
         SessionRecipe.get_instance().get_claims_added_by_other_recipes()
     )
     app_info = SessionRecipe.get_instance().app_info
-    issuer = (
-        app_info.api_domain.get_as_string_dangerous()
-        + app_info.api_base_path.get_as_string_dangerous()
+
+    api_domain_resp = await get_api_domain_or_throw_error(
+        api_domain, app_info, user_context
     )
+
+    issuer = api_domain_resp + app_info.api_base_path.get_as_string_dangerous()
 
     final_access_token_payload = {**access_token_payload, "iss": issuer}
 
@@ -488,6 +491,8 @@ async def create_jwt(
     payload: Dict[str, Any],
     validity_seconds: Optional[int] = None,
     use_static_signing_key: Optional[bool] = None,
+    api_domain: Optional[str] = None,
+    issuer_domain: Optional[str] = None,
     user_context: Union[None, Dict[str, Any]] = None,
 ) -> Union[CreateJwtOkResult, CreateJwtResultUnsupportedAlgorithm]:
     if user_context is None:
@@ -495,7 +500,12 @@ async def create_jwt(
     openid_recipe = SessionRecipe.get_instance().openid_recipe
 
     return await openid_recipe.recipe_implementation.create_jwt(
-        payload, validity_seconds, use_static_signing_key, user_context
+        payload,
+        validity_seconds,
+        use_static_signing_key,
+        issuer_domain,
+        api_domain,
+        user_context,
     )
 
 
@@ -507,7 +517,8 @@ async def get_jwks(user_context: Union[None, Dict[str, Any]] = None) -> GetJWKSR
 
 
 async def get_open_id_discovery_configuration(
-    user_context: Union[None, Dict[str, Any]] = None
+    issuer_domain: Optional[str] = None,
+    user_context: Union[None, Dict[str, Any]] = None,
 ) -> GetOpenIdDiscoveryConfigurationResult:
     if user_context is None:
         user_context = {}
@@ -515,7 +526,7 @@ async def get_open_id_discovery_configuration(
 
     return (
         await openid_recipe.recipe_implementation.get_open_id_discovery_configuration(
-            user_context
+            issuer_domain, user_context
         )
     )