+++ supertokens_python/recipe/session/recipe.py
@@ -133,9 +133,6 @@ class SessionRecipe(RecipeModule):
             "session init: cookie_same_site: %s", self.config.cookie_same_site
         )
         log_debug_message(
-            "session init: cookie_secure: %s", str(self.config.cookie_secure)
-        )
-        log_debug_message(
             "session init: refresh_token_path: %s ",
             self.config.refresh_token_path.get_as_string_dangerous(),
         )