+++ supertokens_python/recipe/thirdparty/providers/apple.py
@@ -142,9 +142,11 @@ class Apple(Provider):
         }
         return AccessTokenAPI(self.access_token_api_url, params)
 
-    def get_redirect_uri(self, user_context: Dict[str, Any]) -> Union[None, str]:
+    def get_redirect_uri(
+        self, api_domain: str, user_context: Dict[str, Any]
+    ) -> Union[None, str]:
         app_info = Supertokens.get_instance().app_info
-        redirect_uri = app_info.api_domain.get_as_string_dangerous()
+        redirect_uri = api_domain
         redirect_uri += app_info.api_base_path.get_as_string_dangerous()
         redirect_uri += APPLE_REDIRECT_HANDLER
         return redirect_uri